+++ tools/@aws-cdk/security-guardian/rules/trust_scope_rules.guard
