+++ tools/@aws-cdk/security-guardian/test/templates/CMCMK-Stack.template.json
